requirements.txt
streamlit
pandas
pdfplumber
python-docx
openai
thefuzz
python-dotenv
fpdf2
openpyxl